requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
mysql-connector-python
pydantic[email]
pytest
pytest-cov
requests
python-multipart
python-jose
jinja2
alembic
aiofiles